+++ CHANGELOG.md
@@ -50,6 +50,7 @@
  * status --details to view changes of tasks
  * --force to overwrite changed tasks
  * --ignore to ignore changes in tasks and consider them finished
+ * new BAD state for tasks when UAP caught the error
 
 **additional stuff**
  * updated documentation and resolved sphinx warnings
@@ -59,7 +60,8 @@
  * hisat2 now takes `library_type` option
  * complete merge with https://github.com/yigbt/uap
  * chronological naming for log files
- * move ping files on job error or interruption (#147)
+ * move execution ping files on job error or interruption (#147)
+ * move queue ping files only on success (#147)
  * no hash in output directories
 
 ## 1.1 (20.01.2020)